requirements.txt
pandas
ollama
google-generativeai